@@ -18,45 +18,45 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.12",
-    "@angular/compiler": "^7.2.12",
-    "@angular/core": "^7.2.12",
-    "@angular/forms": "^7.2.12",
-    "@angular/platform-browser": "^7.2.12",
-    "@angular/platform-browser-dynamic": "^7.2.12",
-    "@angular/router": "^7.2.12",
+    "@angular/animations": "^8.2.7",
+    "@angular/common": "^8.2.7",
+    "@angular/compiler": "^8.2.7",
+    "@angular/core": "^8.2.7",
+    "@angular/forms": "^8.2.7",
+    "@angular/platform-browser": "^8.2.7",
+    "@angular/platform-browser-dynamic": "^8.2.7",
+    "@angular/router": "^8.2.7",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "@types/lodash": "^4.14.123",
     "bootstrap": "^4.3.1",
-    "core-js": "^2.6.5",
     "jquery": "^3.3.1",
     "lodash": "^4.17.11",
-    "ngx-bootstrap": "^4.0.0",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.8.26"
+    "ngx-bootstrap": "^5.1.2",
+    "rxjs": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.12",
-    "@angular/language-service": "^7.2.12",
-    "@types/jasmine": "~3.3.12",
+    "@angular-devkit/build-angular": "~0.803.5",
+    "@angular/cli": "^8.3.5",
+    "@angular/compiler-cli": "^8.2.7",
+    "@angular/language-service": "^8.2.7",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~11.13.0",
-    "codelyzer": "~5.0.0",
+    "@types/node": "~12.7.5",
+    "codelyzer": "~5.1.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.2",
     "ts-node": "^8.1.0",
     "tslint": "^5.16.0",
-    "typedoc": "^0.14.2",
-    "typescript": ">=3.2.0 <3.3.0"
+    "typedoc": "^0.15.0",
+    "typescript": "3.5.3"
   }
 }
