@@ -16,8 +16,8 @@ module.exports = function(config) {
 			clearContext: false // leave Jasmine Spec Runner output visible in browser
 		},
 		coverageIstanbulReporter: {
-			dir: require('path').join(__dirname, '../coverage'),
-			reports: ['html', 'lcovonly'],
+			dir: require('path').join(__dirname, './coverage/web'),
+			reports: ['html', 'lcovonly', 'text-summary'],
 			fixWebpackSourcePaths: true
 		},
 		reporters: ['progress', 'kjhtml'],
@@ -27,6 +27,7 @@ module.exports = function(config) {
 		autoWatch: true,
 		usePolling: true,
 		browsers: ['ChromiumHeadless'],
-		singleRun: true // FIXME: VMだとautoWatchが動作しない（？）ため、一旦singleRunで運用
+		singleRun: true, // FIXME: VMだとautoWatchが動作しない（？）ため、一旦singleRunで運用
+		restartOnFileChange: true
 	});
 };
