@@ -16,4 +16,4 @@ if (environment.production) {
 registerLocaleData(localeJa, 'ja');
 
 platformBrowserDynamic().bootstrapModule(AppModule)
-	.catch(err => console.log(err));
+	.catch(err => console.error(err));
